Reorganize multiple modules into 'app' package
The main.py and various other modules have been restructured into an 'app' package for better organization. This also includes renaming a class in models.py and integrating a new router for 'customers' in main.py. The restructuring aims to improve project maintainability and enhance the working with FastAPI-based application.